+++ turbo/apps/api/src/signals/services/zero-feishu-dispatch.service.ts
@@ -31,7 +31,7 @@ import { safeJsonParse, tapError } from "../utils";
 import { buildFeishuConnectUrl } from "./feishu-connect-token";
 import { publishCustomConnectorUserInvalidationAfterCommit } from "./connector-client-invalidation.service";
 import { disconnectFeishuCustomConnectorOAuthConnection } from "./feishu-custom-connector.service";
-import { publishFeishuOrgChanged } from "./zero-feishu-realtime.service";
+import { publishFeishuOrgChanged } from "./feishu-realtime.service";
 import { listOrgModelPolicies$ } from "./zero-model-policy.service";
 import {
   updateUserModelPreference$,
